@@ -1,25 +1,27 @@
 name: code_builder
-version: 3.7.0
+version: 3.8.0-nullsafety01
+  # Maybe the correct version should be 4.0.0 but it will block build_runner 
+  # because of the code_builder version constraints.
 
 description: >-
   A fluent, builder-based library for generating valid Dart code
 homepage: https://github.com/dart-lang/code_builder
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  built_collection: '>=3.0.0 <6.0.0'
-  built_value: '>=7.0.0 <9.0.0'
-  collection: ^1.14.0
-  matcher: ^0.12.0
-  meta: ^1.0.5
+  built_collection: '>=5.0.0 <6.0.0'
+  built_value: '>=8.0.4 <9.0.0'
+  collection: ^1.15.0
+  matcher: ^0.12.10
+  meta: ^1.3.0
 
 dev_dependencies:
-  build: ^1.0.0
-  build_runner: ^1.1.0
-  built_value_generator: ^7.0.0
-  dart_style: ^1.0.0
-  pedantic: ^1.0.0
-  source_gen: ^0.9.0
-  test: ^1.3.0
+  build: ^2.0.0
+  build_runner: ^1.12.2
+  built_value_generator: ^8.0.4
+  dart_style: ^2.0.0
+  pedantic: ^1.11.0
+  source_gen: ^1.0.0
+  test: ^1.16.8
