@@ -1,4 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// @dart=2.12
 
 part of 'method.dart';
 
@@ -18,61 +19,49 @@ class _$Method extends Method {
   @override
   final BuiltList<Parameter> requiredParameters;
   @override
-  final Code body;
+  final Code? body;
   @override
   final bool external;
   @override
-  final bool lambda;
+  final bool? lambda;
   @override
   final bool static;
   @override
-  final String name;
+  final String? name;
   @override
-  final MethodType type;
+  final MethodType? type;
   @override
-  final MethodModifier modifier;
+  final MethodModifier? modifier;
   @override
-  final Reference returns;
+  final Reference? returns;
 
-  factory _$Method([void Function(MethodBuilder) updates]) =>
+  factory _$Method([void Function(MethodBuilder)? updates]) =>
       (new MethodBuilder()..update(updates)).build() as _$Method;
 
   _$Method._(
-      {this.annotations,
-      this.docs,
-      this.types,
-      this.optionalParameters,
-      this.requiredParameters,
+      {required this.annotations,
+      required this.docs,
+      required this.types,
+      required this.optionalParameters,
+      required this.requiredParameters,
       this.body,
-      this.external,
+      required this.external,
       this.lambda,
-      this.static,
+      required this.static,
       this.name,
       this.type,
       this.modifier,
       this.returns})
       : super._() {
-    if (annotations == null) {
-      throw new BuiltValueNullFieldError('Method', 'annotations');
-    }
-    if (docs == null) {
-      throw new BuiltValueNullFieldError('Method', 'docs');
-    }
-    if (types == null) {
-      throw new BuiltValueNullFieldError('Method', 'types');
-    }
-    if (optionalParameters == null) {
-      throw new BuiltValueNullFieldError('Method', 'optionalParameters');
-    }
-    if (requiredParameters == null) {
-      throw new BuiltValueNullFieldError('Method', 'requiredParameters');
-    }
-    if (external == null) {
-      throw new BuiltValueNullFieldError('Method', 'external');
-    }
-    if (static == null) {
-      throw new BuiltValueNullFieldError('Method', 'static');
-    }
+    BuiltValueNullFieldError.checkNotNull(annotations, 'Method', 'annotations');
+    BuiltValueNullFieldError.checkNotNull(docs, 'Method', 'docs');
+    BuiltValueNullFieldError.checkNotNull(types, 'Method', 'types');
+    BuiltValueNullFieldError.checkNotNull(
+        optionalParameters, 'Method', 'optionalParameters');
+    BuiltValueNullFieldError.checkNotNull(
+        requiredParameters, 'Method', 'requiredParameters');
+    BuiltValueNullFieldError.checkNotNull(external, 'Method', 'external');
+    BuiltValueNullFieldError.checkNotNull(static, 'Method', 'static');
   }
 
   @override
@@ -152,12 +141,12 @@ class _$Method extends Method {
 }
 
 class _$MethodBuilder extends MethodBuilder {
-  _$Method _$v;
+  _$Method? _$v;
 
   @override
   ListBuilder<Expression> get annotations {
     _$this;
-    return super.annotations ??= new ListBuilder<Expression>();
+    return super.annotations;
   }
 
   @override
@@ -169,7 +158,7 @@ class _$MethodBuilder extends MethodBuilder {
   @override
   ListBuilder<String> get docs {
     _$this;
-    return super.docs ??= new ListBuilder<String>();
+    return super.docs;
   }
 
   @override
@@ -181,7 +170,7 @@ class _$MethodBuilder extends MethodBuilder {
   @override
   ListBuilder<Reference> get types {
     _$this;
-    return super.types ??= new ListBuilder<Reference>();
+    return super.types;
   }
 
   @override
@@ -193,7 +182,7 @@ class _$MethodBuilder extends MethodBuilder {
   @override
   ListBuilder<Parameter> get optionalParameters {
     _$this;
-    return super.optionalParameters ??= new ListBuilder<Parameter>();
+    return super.optionalParameters;
   }
 
   @override
@@ -205,7 +194,7 @@ class _$MethodBuilder extends MethodBuilder {
   @override
   ListBuilder<Parameter> get requiredParameters {
     _$this;
-    return super.requiredParameters ??= new ListBuilder<Parameter>();
+    return super.requiredParameters;
   }
 
   @override
@@ -215,13 +204,13 @@ class _$MethodBuilder extends MethodBuilder {
   }
 
   @override
-  Code get body {
+  Code? get body {
     _$this;
     return super.body;
   }
 
   @override
-  set body(Code body) {
+  set body(Code? body) {
     _$this;
     super.body = body;
   }
@@ -239,13 +228,13 @@ class _$MethodBuilder extends MethodBuilder {
   }
 
   @override
-  bool get lambda {
+  bool? get lambda {
     _$this;
     return super.lambda;
   }
 
   @override
-  set lambda(bool lambda) {
+  set lambda(bool? lambda) {
     _$this;
     super.lambda = lambda;
   }
@@ -263,49 +252,49 @@ class _$MethodBuilder extends MethodBuilder {
   }
 
   @override
-  String get name {
+  String? get name {
     _$this;
     return super.name;
   }
 
   @override
-  set name(String name) {
+  set name(String? name) {
     _$this;
     super.name = name;
   }
 
   @override
-  MethodType get type {
+  MethodType? get type {
     _$this;
     return super.type;
   }
 
   @override
-  set type(MethodType type) {
+  set type(MethodType? type) {
     _$this;
     super.type = type;
   }
 
   @override
-  MethodModifier get modifier {
+  MethodModifier? get modifier {
     _$this;
     return super.modifier;
   }
 
   @override
-  set modifier(MethodModifier modifier) {
+  set modifier(MethodModifier? modifier) {
     _$this;
     super.modifier = modifier;
   }
 
   @override
-  Reference get returns {
+  Reference? get returns {
     _$this;
     return super.returns;
   }
 
   @override
-  set returns(Reference returns) {
+  set returns(Reference? returns) {
     _$this;
     super.returns = returns;
   }
@@ -313,20 +302,21 @@ class _$MethodBuilder extends MethodBuilder {
   _$MethodBuilder() : super._();
 
   MethodBuilder get _$this {
-    if (_$v != null) {
-      super.annotations = _$v.annotations?.toBuilder();
-      super.docs = _$v.docs?.toBuilder();
-      super.types = _$v.types?.toBuilder();
-      super.optionalParameters = _$v.optionalParameters?.toBuilder();
-      super.requiredParameters = _$v.requiredParameters?.toBuilder();
-      super.body = _$v.body;
-      super.external = _$v.external;
-      super.lambda = _$v.lambda;
-      super.static = _$v.static;
-      super.name = _$v.name;
-      super.type = _$v.type;
-      super.modifier = _$v.modifier;
-      super.returns = _$v.returns;
+    final $v = _$v;
+    if ($v != null) {
+      super.annotations = $v.annotations.toBuilder();
+      super.docs = $v.docs.toBuilder();
+      super.types = $v.types.toBuilder();
+      super.optionalParameters = $v.optionalParameters.toBuilder();
+      super.requiredParameters = $v.requiredParameters.toBuilder();
+      super.body = $v.body;
+      super.external = $v.external;
+      super.lambda = $v.lambda;
+      super.static = $v.static;
+      super.name = $v.name;
+      super.type = $v.type;
+      super.modifier = $v.modifier;
+      super.returns = $v.returns;
       _$v = null;
     }
     return this;
@@ -334,14 +324,12 @@ class _$MethodBuilder extends MethodBuilder {
 
   @override
   void replace(Method other) {
-    if (other == null) {
-      throw new ArgumentError.notNull('other');
-    }
+    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$Method;
   }
 
   @override
-  void update(void Function(MethodBuilder) updates) {
+  void update(void Function(MethodBuilder)? updates) {
     if (updates != null) updates(this);
   }
 
@@ -357,15 +345,17 @@ class _$MethodBuilder extends MethodBuilder {
               optionalParameters: optionalParameters.build(),
               requiredParameters: requiredParameters.build(),
               body: body,
-              external: external,
+              external: BuiltValueNullFieldError.checkNotNull(
+                  external, 'Method', 'external'),
               lambda: lambda,
-              static: static,
+              static: BuiltValueNullFieldError.checkNotNull(
+                  static, 'Method', 'static'),
               name: name,
               type: type,
               modifier: modifier,
               returns: returns);
     } catch (_) {
-      String _$failedField;
+      late String _$failedField;
       try {
         _$failedField = 'annotations';
         annotations.build();
@@ -390,7 +380,7 @@ class _$MethodBuilder extends MethodBuilder {
 
 class _$Parameter extends Parameter {
   @override
-  final Code defaultTo;
+  final Code? defaultTo;
   @override
   final String name;
   @override
@@ -404,51 +394,36 @@ class _$Parameter extends Parameter {
   @override
   final BuiltList<Reference> types;
   @override
-  final Reference type;
+  final Reference? type;
   @override
   final bool required;
   @override
   final bool covariant;
 
-  factory _$Parameter([void Function(ParameterBuilder) updates]) =>
+  factory _$Parameter([void Function(ParameterBuilder)? updates]) =>
       (new ParameterBuilder()..update(updates)).build() as _$Parameter;
 
   _$Parameter._(
       {this.defaultTo,
-      this.name,
-      this.named,
-      this.toThis,
-      this.annotations,
-      this.docs,
-      this.types,
+      required this.name,
+      required this.named,
+      required this.toThis,
+      required this.annotations,
+      required this.docs,
+      required this.types,
       this.type,
-      this.required,
-      this.covariant})
+      required this.required,
+      required this.covariant})
       : super._() {
-    if (name == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'name');
-    }
-    if (named == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'named');
-    }
-    if (toThis == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'toThis');
-    }
-    if (annotations == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'annotations');
-    }
-    if (docs == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'docs');
-    }
-    if (types == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'types');
-    }
-    if (required == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'required');
-    }
-    if (covariant == null) {
-      throw new BuiltValueNullFieldError('Parameter', 'covariant');
-    }
+    BuiltValueNullFieldError.checkNotNull(name, 'Parameter', 'name');
+    BuiltValueNullFieldError.checkNotNull(named, 'Parameter', 'named');
+    BuiltValueNullFieldError.checkNotNull(toThis, 'Parameter', 'toThis');
+    BuiltValueNullFieldError.checkNotNull(
+        annotations, 'Parameter', 'annotations');
+    BuiltValueNullFieldError.checkNotNull(docs, 'Parameter', 'docs');
+    BuiltValueNullFieldError.checkNotNull(types, 'Parameter', 'types');
+    BuiltValueNullFieldError.checkNotNull(required, 'Parameter', 'required');
+    BuiltValueNullFieldError.checkNotNull(covariant, 'Parameter', 'covariant');
   }
 
   @override
@@ -514,16 +489,16 @@ class _$Parameter extends Parameter {
 }
 
 class _$ParameterBuilder extends ParameterBuilder {
-  _$Parameter _$v;
+  _$Parameter? _$v;
 
   @override
-  Code get defaultTo {
+  Code? get defaultTo {
     _$this;
     return super.defaultTo;
   }
 
   @override
-  set defaultTo(Code defaultTo) {
+  set defaultTo(Code? defaultTo) {
     _$this;
     super.defaultTo = defaultTo;
   }
@@ -567,7 +542,7 @@ class _$ParameterBuilder extends ParameterBuilder {
   @override
   ListBuilder<Expression> get annotations {
     _$this;
-    return super.annotations ??= new ListBuilder<Expression>();
+    return super.annotations;
   }
 
   @override
@@ -579,7 +554,7 @@ class _$ParameterBuilder extends ParameterBuilder {
   @override
   ListBuilder<String> get docs {
     _$this;
-    return super.docs ??= new ListBuilder<String>();
+    return super.docs;
   }
 
   @override
@@ -591,7 +566,7 @@ class _$ParameterBuilder extends ParameterBuilder {
   @override
   ListBuilder<Reference> get types {
     _$this;
-    return super.types ??= new ListBuilder<Reference>();
+    return super.types;
   }
 
   @override
@@ -601,13 +576,13 @@ class _$ParameterBuilder extends ParameterBuilder {
   }
 
   @override
-  Reference get type {
+  Reference? get type {
     _$this;
     return super.type;
   }
 
   @override
-  set type(Reference type) {
+  set type(Reference? type) {
     _$this;
     super.type = type;
   }
@@ -639,17 +614,18 @@ class _$ParameterBuilder extends ParameterBuilder {
   _$ParameterBuilder() : super._();
 
   ParameterBuilder get _$this {
-    if (_$v != null) {
-      super.defaultTo = _$v.defaultTo;
-      super.name = _$v.name;
-      super.named = _$v.named;
-      super.toThis = _$v.toThis;
-      super.annotations = _$v.annotations?.toBuilder();
-      super.docs = _$v.docs?.toBuilder();
-      super.types = _$v.types?.toBuilder();
-      super.type = _$v.type;
-      super.required = _$v.required;
-      super.covariant = _$v.covariant;
+    final $v = _$v;
+    if ($v != null) {
+      super.defaultTo = $v.defaultTo;
+      super.name = $v.name;
+      super.named = $v.named;
+      super.toThis = $v.toThis;
+      super.annotations = $v.annotations.toBuilder();
+      super.docs = $v.docs.toBuilder();
+      super.types = $v.types.toBuilder();
+      super.type = $v.type;
+      super.required = $v.required;
+      super.covariant = $v.covariant;
       _$v = null;
     }
     return this;
@@ -657,14 +633,12 @@ class _$ParameterBuilder extends ParameterBuilder {
 
   @override
   void replace(Parameter other) {
-    if (other == null) {
-      throw new ArgumentError.notNull('other');
-    }
+    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$Parameter;
   }
 
   @override
-  void update(void Function(ParameterBuilder) updates) {
+  void update(void Function(ParameterBuilder)? updates) {
     if (updates != null) updates(this);
   }
 
@@ -675,17 +649,22 @@ class _$ParameterBuilder extends ParameterBuilder {
       _$result = _$v ??
           new _$Parameter._(
               defaultTo: defaultTo,
-              name: name,
-              named: named,
-              toThis: toThis,
+              name: BuiltValueNullFieldError.checkNotNull(
+                  name, 'Parameter', 'name'),
+              named: BuiltValueNullFieldError.checkNotNull(
+                  named, 'Parameter', 'named'),
+              toThis: BuiltValueNullFieldError.checkNotNull(
+                  toThis, 'Parameter', 'toThis'),
               annotations: annotations.build(),
               docs: docs.build(),
               types: types.build(),
               type: type,
-              required: required,
-              covariant: covariant);
+              required: BuiltValueNullFieldError.checkNotNull(
+                  required, 'Parameter', 'required'),
+              covariant: BuiltValueNullFieldError.checkNotNull(
+                  covariant, 'Parameter', 'covariant'));
     } catch (_) {
-      String _$failedField;
+      late String _$failedField;
       try {
         _$failedField = 'annotations';
         annotations.build();
