@@ -1,6 +1,7 @@
 // Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+// @dart=2.12
 
 import 'package:code_builder/code_builder.dart';
 import 'package:test/test.dart';
@@ -22,7 +23,7 @@ void main() {
   });
 
   group('in a Null Safety library', () {
-    DartEmitter emitter;
+    late DartEmitter emitter;
 
     setUp(() => emitter = DartEmitter.scoped(useNullSafetySyntax: true));
 
