@@ -1,11 +1,12 @@
 // Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+// @dart=2.12
 
 import 'visitors.dart';
 
 abstract class Spec {
-  R accept<R>(SpecVisitor<R> visitor, [R context]);
+  R accept<R>(SpecVisitor<R> visitor, [R? context]);
 }
 
 /// Returns a generic [Spec] that is lazily generated when visited.
@@ -17,6 +18,6 @@ class _LazySpec implements Spec {
   const _LazySpec(this.generate);
 
   @override
-  R accept<R>(SpecVisitor<R> visitor, [R context]) =>
+  R accept<R>(SpecVisitor<R> visitor, [R? context]) =>
       generate().accept(visitor, context);
 }
