@@ -1,6 +1,7 @@
 // Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+// @dart=2.12
 
 import 'package:built_collection/built_collection.dart';
 
@@ -15,5 +16,5 @@ abstract class HasAnnotations {
 /// Compliment to the [HasAnnotations] mixin for metadata [annotations].
 abstract class HasAnnotationsBuilder {
   /// Annotations as metadata on the node.
-  ListBuilder<Expression> annotations;
+  abstract ListBuilder<Expression> annotations;
 }
