@@ -1,6 +1,7 @@
 // Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+// @dart=2.12
 
 import 'package:built_collection/built_collection.dart';
 
@@ -13,5 +14,5 @@ abstract class HasGenerics {
 
 abstract class HasGenericsBuilder {
   /// Generic type parameters.
-  ListBuilder<Reference> types;
+  abstract ListBuilder<Reference> types;
 }
